+++ library/SpatialDomains/MeshPartitionScotch.cpp
@@ -62,7 +62,7 @@ MeshPartitionScotch::~MeshPartitionScotch()
 {
 }
 
-void MeshPartitionScotch::PartitionGraphImpl(
+void MeshPartitionScotch::v_PartitionGraphImpl(
     int &nVerts, int &nVertConds, Nektar::Array<Nektar::OneD, int> &xadj,
     Nektar::Array<Nektar::OneD, int> &adjcy,
     Nektar::Array<Nektar::OneD, int> &vertWgt,